fix: upload image santri,  slug berita

+++ app/Models/Santri_model.php
@@ -137,7 +137,7 @@ class Santri_model extends Model
     {
         return  $this->where('id !=', $id)->where("nis", $nis)->first();
     }
-    
+
     // update status santri
     public function updateDatas($id, $data)
     {
@@ -206,7 +206,7 @@ class Santri_model extends Model
         // update
         $this->update($id, $data);
         //foto diri
-        if (is_uploaded_file($data["foto_diri"])) {
+        if (isset($data["foto_diri"]) &&  is_uploaded_file($data["foto_diri"])) {
             $imageFile = $data["foto_diri"];
             $nameFileFotoDiri = $data['foto_diri']->getRandomName();
             $data['image'] = $nameFileFotoDiri;
@@ -217,7 +217,7 @@ class Santri_model extends Model
             );
         }
         //foto kk
-        if (is_uploaded_file($data["foto_kk"])) {
+        if (isset($data["foto_kk"]) && is_uploaded_file($data["foto_kk"])) {
             $kkFile = $data["foto_kk"];
             $nameFileFotoKK = $data['foto_kk']->getRandomName();
             $data['file_kk'] = $nameFileFotoKK;
@@ -228,7 +228,7 @@ class Santri_model extends Model
             );
         }
         //foto akte
-        if (is_uploaded_file($data["foto_akte"])) {
+        if (isset($data["foto_akte"]) &&  is_uploaded_file($data["foto_akte"])) {
             $akteFile = $data["foto_akte"];
             $nameFileFotoakte = $data['foto_akte']->getRandomName();
             $data['file_akte'] = $nameFileFotoakte;
@@ -240,7 +240,7 @@ class Santri_model extends Model
         }
 
         //foto ijazah
-        if (is_uploaded_file($data["foto_ijazah"])) {
+        if (isset($data["foto_ijazah"]) && is_uploaded_file($data["foto_ijazah"])) {
             $ijazahFile = $data["foto_ijazah"];
             $nameFileFotoijazah = $data['foto_ijazah']->getRandomName();
             $data['file_ijazah'] = $nameFileFotoijazah;
@@ -252,7 +252,7 @@ class Santri_model extends Model
         }
 
         // foto skhu
-        if (is_uploaded_file($data["foto_skhu"])) {
+        if (isset($data["foto_skhu"]) && is_uploaded_file($data["foto_skhu"])) {
             $skhuFile = $data["foto_skhu"];
             $nameFileFotoskhu = $data['foto_skhu']->getRandomName();
             $data['file_skhu'] = $nameFileFotoskhu;
@@ -266,13 +266,13 @@ class Santri_model extends Model
 
     public function deleteData($id)
     {
-        $pathDir = "../public/upload/$id/";
+        $pathDir = "../public/upload/santri/$id/";
         $image = "../public/upload/santri/$id/" . $this->find($id)["image"];
         $file_kk = "../public/upload/santri/$id/" . $this->find($id)["file_kk"];
         $file_akte = "../public/upload/santri/$id/" . $this->find($id)["file_akte"];
         $file_ijazah = "../public/upload/santri/$id/" . $this->find($id)["file_ijazah"];
         $file_skhu = "../public/upload/santri/$id/" . $this->find($id)["file_skhu"];
-        if (file_exists($image, $file_kk, $file_akte, $file_ijazah, $file_skhu)) {
+        if (file_exists($image) && file_exists($file_kk) && file_exists($file_akte) && file_exists($file_ijazah) && file_exists($file_skhu)) {
             unlink($image);
             unlink($file_kk);
             unlink($file_akte);
